Adds garlic, and tweaks bloodsuckers a bunch.

+++ code/modules/antagonists/bloodsucker/items/bloodsucker_organs.dm
@@ -5,7 +5,8 @@
 	var/obj/item/organ/O
 	// Heart
 	O = owner.current.getorganslot(ORGAN_SLOT_HEART)
-	if(!istype(O, /obj/item/organ/heart/vampheart))
+
+	if(!istype(O, /obj/item/organ/heart/vampheart) && !istype(O, /obj/item/organ/heart/demon))
 		qdel(O)
 		var/obj/item/organ/heart/vampheart/H = new
 		H.Insert(owner.current)
